+++ extensions/ql-vscode/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add new command in query history view to view the log file of a
   query.
 - Request user acknowledgment before updating the CodeQL binaries.
+- Warn when using the deprecated `codeql.cmd` launcher on windows.
 
 ## 1.1.0 - 17 March 2020